+++ test/liberator_mixin/json/core_test.clj
@@ -95,7 +95,36 @@
             (:body response))))))
 
 (deftest with-params-parsed-as-json
-  (testing "parses simple JSON params"
+  #_(testing "parses simple JSON params"
+      (let [resource
+            (->
+              (l/build-resource
+                (json/with-json-media-type)
+                (json/with-params-parsed-as-json)
+                {:allowed-methods
+                 [:get]
+
+                 :handle-ok
+                 (fn [{:keys [request]}]
+                   (let [params (:params request)]
+                     {:correct-params
+                      (and
+                        (= (:thing1 params) [1 2 3])
+                        (= (:thing2 params)
+                          {:key-1 "value"}))}))})
+              (keyword-params/wrap-keyword-params)
+              (params/wrap-params))
+            request (->
+                      (ring/request :get "/")
+                      (ring/header "Content-Type" json/json-media-type)
+                      (ring/query-string
+                        "thing1=[1,2,3]&thing2={\"key1\":\"value\"}"))
+            response (call-resource resource request)]
+        (is (= 200 (:status response)))
+        (is (= {:correct-params true}
+              (:body response)))))
+
+  (testing "parses multi-valued JSON params"
     (let [resource
           (->
             (l/build-resource
@@ -108,17 +137,16 @@
                (fn [{:keys [request]}]
                  (let [params (:params request)]
                    {:correct-params
-                    (and
-                      (= (:thing1 params) [1 2 3])
-                      (= (:thing2 params)
-                        {:key-1 "value"}))}))})
+                    (= (:filter params) [["firstName", "Aaron"]
+                                         ["lastName", "Cross"]])}))})
             (keyword-params/wrap-keyword-params)
             (params/wrap-params))
           request (->
                     (ring/request :get "/")
                     (ring/header "Content-Type" json/json-media-type)
                     (ring/query-string
-                      "thing1=[1,2,3]&thing2={\"key1\":\"value\"}"))
+                      (str "filter=[\"firstName\",\"Aaron\"]&"
+                        "filter=[\"lastName\",\"Cross\"]")))
           response (call-resource resource request)]
       (is (= 200 (:status response)))
       (is (= {:correct-params true}